fix: 관리자 구매인증 조회 엔드포인트 변경
